Remoção da classe base ConselhoClasseTesteBase
