+++ teste/SME.SGP.TesteIntegracao/ConselhoDeClasse/Ao_gerar_parecer_conclusivo.cs
@@ -11,7 +11,6 @@ using SME.SGP.Aplicacao.Interfaces;
 using SME.SGP.Dominio;
 using SME.SGP.Infra;
 using SME.SGP.Infra.Dtos;
-using SME.SGP.TesteIntegracao.ConselhoDeClasse.Base;
 using SME.SGP.TesteIntegracao.ConselhoDeClasse.ServicosFakes;
 using SME.SGP.TesteIntegracao.ServicosFakes;
 using SME.SGP.TesteIntegracao.Setup;
