+++ teste/SME.SGP.TesteIntegracao/ConselhoDeClasse/Ao_reprocessar_conselho_classe_aluno.cs
@@ -7,7 +7,6 @@ using Shouldly;
 using SME.SGP.Aplicacao;
 using SME.SGP.Dominio;
 using SME.SGP.Infra;
-using SME.SGP.TesteIntegracao.ConselhoDeClasse.Base;
 using SME.SGP.TesteIntegracao.ConselhoDeClasse.ServicosFakes;
 using SME.SGP.TesteIntegracao.Setup;
 using Xunit;
